simplify main logic, move configs to file load instead of inside functions
